pyproject.toml
[project]
name = "civtai-cli-dl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "about-time==4.2.1",
    "alive-progress==3.3.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "graphemeu==0.7.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "python-dotenv==1.1.1",
    "rich>=14.1.0",
    "sniffio==1.3.1",
    "typing-extensions==4.14.1",
]
